requirements.txt
#先把pip版本降到24.1以下
pip==24.0
tensorboard==2.4.1
matplotlib==3.4.1
transformers==4.26.0
jieba==0.42.1
scikit-learn==0.24.1
seqeval==1.2.2
opt-einsum==3.3.0
wandb==0.12.7
ujson==5.6.0
huggingface_hub==0.11.0
tensorboardX==2.5.1
nltk==3.8
protobuf==3.20.1
numpy==1.21.0
ipdb==0.13.11
pytorch-crf==0.7.2
tqdm==4.66.1
openai==0.28.0
Jinja2==3.1.2
datasets==2.13.2
pyhocon==0.3.60
torch>=1.5,<=1.11
pandas>=1.3.0
py2neo>=2021.2.3
hydra-core>=1.1.0
zhconv>=1.4.0
paddlenlp>=2.4.0
paddlepaddle>=2.4.0
rouge>=1.0.0
prettytable>=2.0.0
pyarrow>=5.0.0
xlrd2>=1.3.0
docx>=0.2.4
bmtrain>=0.2.0
deepspeed>=0.7.0
nlpaug>=1.1.0
tiktoken>=0.3.0
